Move adapters to kv
