+++ src/raw/mod.rs
@@ -53,3 +53,5 @@ pub use wrappers::apply_wrapper;
 
 mod ops;
 pub use ops::*;
+
+pub mod adapters;
