+++ src/services/memory/backend.rs
@@ -18,7 +18,7 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use parking_lot::Mutex;
 
-use crate::adapters::kv;
+use crate::raw::adapters::kv;
 use crate::raw::*;
 use crate::Result;
 use crate::Scheme;
