+++ src/services/rocksdb/backend.rs
@@ -19,7 +19,7 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use rocksdb::TransactionDB;
 
-use crate::adapters::kv;
+use crate::raw::adapters::kv;
 use crate::raw::*;
 use crate::Result;
 use crate::*;
